Completed parsing of trade log models.

+++ MPL.X4.Library/SaveGame/Data/IGameData.cs
@@ -13,5 +13,5 @@ public interface IGameData
     /// <summary>
     /// Gets the game time.
     /// </summary>
-    decimal Time { get; }
+    double Time { get; }
 }
