+++ MPL.X4.Services/SaveGame/Data/Parser/TradeLogEntryDataParser.cs
@@ -31,7 +31,7 @@ internal class TradeLogEntryDataParser(
         {
             BuyerId = buyer,
             SellerId = seller,
-            Time = (decimal)time.Value,
+            Time = time.Value,
             Price = price.Value,
             Volume = volume.Value,
             Ware = ware
